refactor(gateway): split the responses modules and share the route preamble
